requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
Django==4.1.7
django-admin-inline-paginator==0.4.0
django-axes==5.40.1
django-cors-headers==3.14.0
django-elasticsearch-dsl==7.3
django-encrypted-model-fields==0.6.5
django-ipware==4.0.2
django-model-utils==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
elastic-apm==6.15.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
idna==3.4
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
packaging==23.0
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-logstash==0.4.8
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.14
wrapt==1.15.0
